+++ substra/sdk/schemas.py
@@ -58,6 +58,12 @@ ALGO_INPUTS_PER_CATEGORY = {
         "opener": {"kind": "ASSET_DATA_MANAGER", "multiple": False, "optional": False},
         "predictions": {"kind": "ASSET_MODEL", "multiple": False, "optional": False},
     },
+    "ALGO_PREDICT": {
+        "datasamples": {"kind": "ASSET_DATA_SAMPLE", "multiple": True, "optional": False},
+        "opener": {"kind": "ASSET_DATA_MANAGER", "multiple": False, "optional": False},
+        "model": {"kind": "ASSET_MODEL", "multiple": False, "optional": False},
+        "shared": {"kind": "ASSET_MODEL", "multiple": False, "optional": True},
+    },
 }
 
 ALGO_OUTPUTS_PER_CATEGORY = {
@@ -74,6 +80,7 @@ ALGO_OUTPUTS_PER_CATEGORY = {
     "ALGO_METRIC": {
         "performance": {"kind": "ASSET_PERFORMANCE", "multiple": False},
     },
+    "ALGO_PREDICT": {"predictions": {"kind": "ASSET_MODEL", "multiple": False}},
 }
 
 
@@ -107,6 +114,7 @@ class AlgoCategory(str, enum.Enum):
     composite = "ALGO_COMPOSITE"
     aggregate = "ALGO_AGGREGATE"
     metric = "ALGO_METRIC"
+    predict = "ALGO_PREDICT"
 
 
 class _PydanticConfig(pydantic.BaseModel):